fix to switch to double

+++ rSYBL-control-service-pom/rSYBL-sybl-processing-unit/src/main/java/at/ac/tuwien/dsg/sybl/syblProcessingUnit/processing/Utils.java
@@ -710,7 +710,7 @@ public  boolean evaluateCondition(String condition) throws NoSuchMethodException
 	}
 	//SYBLDirectivesEnforcementLogger.logger.info("Evaluating condition " +s[1]+" and terms "+s[0]+" and "+s[2]);
 	if (s[1].equalsIgnoreCase("<")){
-		if (((Float)evaluateTerm(s[0]))<((Float)evaluateTerm(s[2])))return true;
+		if (((Double)evaluateTerm(s[0]))<((Double)evaluateTerm(s[2])))return true;
 		else return false;
 	}
 	switch (s[1]) {
